Convert wmfjs to typescript
