+++ package.json
@@ -8,7 +8,7 @@
         "fix": "eclint fix src/* src/**/* test/*.js",
         "clean": "rimraf dist",
         "compile": "tsc",
-        "bundle": "rollup --config",
+        "bundle": "rollup --config rollup.config.rtfjs.js && rollup --config rollup.config.wmfjs.js",
         "build": "npm run clean && npm run compile && npm run bundle",
         "generate-testcase": "node test/generate-testcase.js",
         "test": "mocha test/test-rtfjs.js"
