+++ samples/01_rtf/rtf.html
@@ -9,7 +9,7 @@
     <script src="../.common/dep/jquery.svgfilter.min.js"></script>
 
     <script src="../../dist/rtf.bundle.js"></script>
-    <script src="../../dist/wmf.js"></script>
+    <script src="../../dist/wmf.bundle.js"></script>
     <script src="../../dist/emf.js"></script>
 
     <link href="../.common/css/jquery.svg.css" rel="stylesheet" />
